API restructuring. Support for querying active flows
